fix compile on win64

+++ Common/ListUtils.cpp
@@ -12,6 +12,7 @@
 #include <string.h>
 #include <errno.h>
 #include <sys/types.h>
+#include <inttypes.h>
 #include "MallocUtils.h"
 #include "ListUtils.h"
 #include "TreeUtils.h"
@@ -207,7 +208,7 @@ int List_ISearch(List_T * liste, void *data,
   ptr = (void *)bsearch(data, liste->array, liste->n, liste->size, fcmp);
   if(ptr == NULL)
     return (-1);
-  return (((long)ptr - (long)liste->array) / liste->size);
+  return (((intptr_t)ptr - (intptr_t)liste->array) / liste->size);
 }
 
 int List_ISearchSeq(List_T * liste, void *data,
@@ -315,7 +316,7 @@ int List_Suppress(List_T *liste, void *data,
   if (ptr == NULL) return(0);
   
   liste->n--;
-  len = liste->n - (((long)ptr - (long)liste->array) / liste->size);
+  len = liste->n - (((intptr_t)ptr - (intptr_t)liste->array) / liste->size);
   if (len > 0) memmove(ptr, ptr + liste->size, len * liste->size);
   return(1);
 }
@@ -330,7 +331,7 @@ int List_PSuppress(List_T * liste, int index)
     return (0);
 
   liste->n--;
-  len = liste->n - (((long)ptr - (long)liste->array) / liste->size);
+  len = liste->n - (((intptr_t)ptr - (intptr_t)liste->array) / liste->size);
   if(len > 0)
     memmove(ptr, ptr + liste->size, len * liste->size);
   return (1);
