+++ Common/avl.cpp
@@ -33,6 +33,7 @@
 // Modified for Gmsh (C++, 64 bits, ...)
 
 #include <stdio.h>
+#include <inttypes.h>
 
 #include "avl.h"
 #include "MallocUtils.h"
@@ -50,7 +51,7 @@
 
 #define COMPARE(key, nodekey, compare)                  \
     ((compare == avl_numcmp) ?                          \
-        (long int) key - (long int) nodekey :                   \
+        (intptr_t) key - (intptr_t) nodekey :                   \
         (*compare)(key, nodekey))
 
 static void avl_record_gen_forward(avl_node *node, avl_generator *gen);
@@ -362,7 +363,7 @@ static avl_node *new_node(void *key, void *value)
 }
 int avl_numcmp(const void *x, const void*y)
 {
-    return (long int) x - (long int) y;
+    return (intptr_t) x - (intptr_t) y;
 }
 
 int avl_check_tree(avl_tree *tree)
